+++ GoogleDFP/Android/SampleApp/app/build.gradle
@@ -19,7 +19,7 @@ android {
 }
 
 dependencies {
-    compile files('libs/tealium.4.1.1c.jar')
     compile 'com.android.support:appcompat-v7:22.2.1'
     compile 'com.google.android.gms:play-services-ads:7.5.0'
+    compile files('libs/tealium.4.1.3c.jar')
 }
