+++ GoogleDFP/Android/SampleApp/app/src/main/java/com/tealium/blankapp/SecondActivity.java
@@ -1,0 +1,14 @@
+package com.tealium.blankapp;
+
+import android.app.Activity;
+import android.os.Bundle;
+
+public class SecondActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_second);
+    }
+
+}
