+++ GoogleDFP/Android/SampleApp/app/src/main/java/com/tealium/googledfp/BannerAdIdentifier.java
