+++ GoogleDFP/Android/SampleApp/app/src/main/res/menu/main.xml
