+++ GoogleDFP/Android/SampleApp/app/src/main/res/values-sw600dp/dimens.xml
