+++ GoogleDFP/Android/SampleApp/app/src/main/res/values-sw720dp-land/dimens.xml
