+++ GoogleDFP/Android/SampleApp/app/src/main/res/values/strings.xml
@@ -3,6 +3,7 @@
 
     <string name="app_name">BlankApp</string>
     <string name="action_settings">Settings</string>
-    <string name="hello_world">Hello world!</string>
+    <string name="main_button_text">SecondActivity</string>
+    <string name="title_activity_second">SecondActivity</string>
 
 </resources>
