+++ GoogleDFP/Android/SampleApp/app/src/main/res/values/styles.xml
