+++ GoogleDFP/Android/SampleApp/local.properties
@@ -7,5 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Fri Aug 14 05:59:52 PDT 2015
+#Wed Aug 26 08:32:44 PDT 2015
 sdk.dir=/Library/android-sdk
