+++ GoogleDFP/Android/SampleApp/settings.gradle
